pyincept-37: Refactoring: extract methods for TestTemplateArchetypes to superclass.
